Remove extracted elements from VaccineCard and LabCard components
