+++ src/components/LabCard/LabCard.js
@@ -1,34 +1,27 @@
 import React, { useState, useEffect } from 'react';
 import {
-  Box, Card, CardContent, Grid, Typography, Divider, IconButton, List, ListItem
+  Box, Grid, Typography, Divider, List, ListItem
 } from '@mui/material';
-import VisibilityIcon from '@mui/icons-material/Visibility';
 import { useHealthCardDataContext } from 'components/HealthCardDataProvider';
 import { getPatientData } from 'utils/qrHelpers';
 import { useTranslation } from 'react-i18next';
-import xIcon from 'assets/x-icon.png';
-import useStyles from './styles';
+import useStyles from '../HealthCard/styles';
 import { getLabTestCodeDisplay, getLabTestResultCodeDisplay } from './labValueSets';
 
 const LabCard = () => {
   const styles = useStyles();
   const { t } = useTranslation();
-  const { jws, issuerVerified, issuerDisplayName } = useHealthCardDataContext();
+  const { jws } = useHealthCardDataContext();
   const [patientData, setPatientData] = useState({
     name: '',
     labResults: [],
   });
-  const [showDateOfBirth, setShowDateOfBirth] = useState(false);
 
   useEffect(() => {
     if (jws) {
       setPatientData(getPatientData(jws));
     }
   }, [jws]);
-
-  const toggleShowDateOfBirth = () => {
-    setShowDateOfBirth(!showDateOfBirth);
-  };
 
   const displayLabTestCode = (code) => {
     if (code?.coding && code?.coding.length >= 0) {
@@ -77,13 +70,13 @@ const LabCard = () => {
   }
 
   const HealthCardLabResult = ({ labResult }) => (
-    <Box className={[styles.flexCol, styles.labRoot].join(' ')}>
+    <Box className={[styles.flexCol, styles.vaccinationRoot].join(' ')}>
       <Box className={styles.divider}>
         <Divider className={styles.line2} />
       </Box>
       <Grid container className={styles.grid}>
         <Grid container item className={styles.gridRow} spacing={0}>
-          <Grid item xs={3} sm={2} className={styles.gridLabel}>
+          <Grid item xs={3} sm={2} className={styles.labGridLabel}>
             <Typography>{t('healthcarddisplay.Lab Test')}</Typography>
           </Grid>
           <Grid item xs={9} sm={10} className={styles.gridItem}>
@@ -96,7 +89,7 @@ const LabCard = () => {
         </Grid>
 
         <Grid container item className={styles.gridRow} spacing={0}>
-          <Grid item xs={3} sm={2} className={styles.gridLabel}>
+          <Grid item xs={3} sm={2} className={styles.labGridLabel}>
             <Typography>{t('healthcarddisplay.Date')}</Typography>
           </Grid>
           <Grid item xs={9} sm={10} className={styles.gridItem}>
@@ -107,7 +100,7 @@ const LabCard = () => {
         </Grid>
 
         <Grid container item className={styles.gridRow} spacing={0}>
-          <Grid item xs={3} sm={2} className={styles.gridLabel}>
+          <Grid item xs={3} sm={2} className={styles.labGridLabel}>
             <Typography>{t('healthcarddisplay.Lab Performer')}</Typography>
           </Grid>
           <Grid item xs={9} sm={10} className={styles.gridItem}>
@@ -118,7 +111,7 @@ const LabCard = () => {
         </Grid>
 
         <Grid container item className={styles.gridRow} spacing={0}>
-          <Grid item xs={3} sm={2} className={styles.gridLabel}>
+          <Grid item xs={3} sm={2} className={styles.labGridLabel}>
             <Typography>{t('healthcarddisplay.Test Result')}</Typography>
           </Grid>
           <Grid item xs={9} sm={10} className={styles.gridItem}>
@@ -130,7 +123,7 @@ const LabCard = () => {
 
         {labResult.referenceRange && labResult.referenceRange.length > 0 && (
           <Grid container item className={styles.gridRow} spacing={0}>
-            <Grid item xs={3} sm={2} className={styles.gridLabel}>
+            <Grid item xs={3} sm={2} className={styles.labGridLabel}>
               <Typography>{t('healthcarddisplay.Reference Range')}</Typography>
             </Grid>
             <Grid item xs={9} sm={10} className={styles.gridItem}>
@@ -144,73 +137,21 @@ const LabCard = () => {
   );
 
   return (
-    <Box display="flex" className={styles.healthCard}>
-      <Card display="flex" className={styles.card}>
-        <CardContent className={styles.cardContent}>
-          <Box className={[styles.flexCol, styles.patientData].join(' ')}>
-            <Typography className={styles.nameLabel}>
-              {t('healthcarddisplay.NAME')}
-            </Typography>
-            <Typography className={styles.name}>{patientData.name}</Typography>
-          </Box>
-          <Box className={[styles.flexCol, styles.patientData].join(' ')}>
-            <Typography className={styles.dateOfBirthLabel}>{t('healthcarddisplay.DATE OF BIRTH')}</Typography>
-            <Box className={styles.dateOfBirthRow}>
-              <Typography className={styles.dateOfBirth} hidden={!showDateOfBirth} id="dateOfBirth">
-                {patientData.dateOfBirth}
-              </Typography>
-              <Typography className={styles.dateOfBirth} hidden={showDateOfBirth}>
-                xxxx-xx-xx
-              </Typography>
-              <IconButton
-                className={styles.eyeOutline}
-                aria-label="toggle datofbirth visibility"
-                onClick={toggleShowDateOfBirth}
-              >
-                <VisibilityIcon />
-              </IconButton>
-            </Box>
-          </Box>
-          <Divider className={styles.line} />
-          <Box className={[styles.flexCol, styles.labRecordBox].join(' ')}>
-            <Box className={styles.labRecordLabel}>
-              <Typography className={styles.labRecordLabelText}>
-                {t('healthcarddisplay.LAB RESULT RECORD')}
-              </Typography>
-            </Box>
-            <List className={styles.labRecordList}>
-              {patientData.labResults.map((item) => (
-                <div key={item.fullUrl}>
-                  <ListItem className={styles.labRecordListItem}>
-                    <HealthCardLabResult labResult={item.resource} />
-                  </ListItem>
-                </div>
-              ))}
-            </List>
-          </Box>
-          <Grid item maxWidth="xs" xs={10.5} justifyContent="left" alignItems="flex-end">
-            <Box display="flex" alignItems="center" justifyContent="left" flexDirection="row">
-              {issuerVerified ? (
-                <Typography className={styles.covid19lab}>{issuerDisplayName}</Typography>
-              ) : (
-                <>
-                  <img
-                    src={xIcon}
-                    alt="Bottom Banner Issuer Icon"
-                    style={{
-                      height: '1.5rem',
-                      marginRight: '1rem'
-                    }}
-                  />
-                  <Typography variant="h6" className={styles.unverifiedText}>
-                    {t('healthcarddisplay.Issuer not recognized')}
-                  </Typography>
-                </>
-              )}
-            </Box>
-          </Grid>
-        </CardContent>
-      </Card>
+    <Box className={[styles.flexCol, styles.vaccinationRecordBox].join(' ')}>
+      <Box className={styles.vaccinationRecordLabel}>
+        <Typography className={styles.vaccinationRecordLabelText}>
+          {t('healthcarddisplay.LAB RESULT RECORD')}
+        </Typography>
+      </Box>
+      <List className={styles.vaccinationRecordList}>
+        {patientData.labResults.map((item) => (
+          <div key={item.fullUrl}>
+            <ListItem className={styles.vaccinationRecordListItem}>
+              <HealthCardLabResult labResult={item.resource} />
+            </ListItem>
+          </div>
+        ))}
+      </List>
     </Box>
   );
 };
